fix: ARM building due to conflicting types for the PMM shim

+++ kernel/arch/arm/integrator-cp/pmm-shim.c
@@ -2,7 +2,7 @@
 #include <stddef.h>
 #include <stdint.h>
 #include <cedille/pmm.h>
-void pmm_shim_alloc_frame(uint32_t address, int kernel, int rw) {
+void pmm_shim_alloc_frame(uintptr_t address, int kernel, int rw) {
 }
-void pmm_shim_free_frame(uint32_t address) {
+void pmm_shim_free_frame(uintptr_t address) {
 }
